+++ libMx/d_net_amplifier.c
@@ -380,7 +380,7 @@ mxd_network_amplifier_get_parameter( MX_AMPLIFIER *amplifier )
 		fname, amplifier->record->name,
 		mx_get_field_label_string( amplifier->record,
 						amplifier->parameter_type ),
-		amplifier->parameter_type));
+		(int) amplifier->parameter_type));
 
 	switch( amplifier->parameter_type ) {
 	case MXLV_AMP_GAIN_RANGE:
